validation pipes generalization
- eliminated validation functions
- dropped validation unit tests
- added validation e2e tests
- generalized both validations and e2e tests
