+++ src/bookings/DTO/create-booking.dto.ts
@@ -1,17 +1,18 @@
 import { IsEmail, IsInt, IsNotEmpty, IsString, Min, MinLength } from 'class-validator';
 
-export class DTO_booking_create {
+  export class DTO_booking_create {
 
-  @IsNotEmpty()
-  @IsInt()
-  showtimeId: number;
+    @IsNotEmpty()
+    @IsInt()
+    @Min(1)
+    showtimeId: number;
 
-  @IsNotEmpty()
-  @Min(1)
-  seatNumber: number;
+    @IsNotEmpty()
+    @Min(1)
+    seatNumber: number;
 
-  @IsNotEmpty()
-  @IsString()
-  userId: string;
-    
-}
+    @IsNotEmpty()
+    @IsString()
+    userId: string;
+      
+  }
