+++ src/bookings/booking.entity.ts
@@ -1,10 +1,17 @@
 import { Entity, Column, PrimaryGeneratedColumn, ManyToOne, JoinColumn } from 'typeorm';
 import { Showtime } from '../showtimes/showtime.entity';
 
+// --------- 2.2.1 requirements 
+
 @Entity()
 export class Booking {
+
+  // --- Primary Key -------------------------------------------
+
   @PrimaryGeneratedColumn('uuid')
   bookingId: string;
+
+  // --- Foreign Key -------------------------------------------
 
   @Column()
   showtimeId: number;
@@ -12,6 +19,8 @@ export class Booking {
   @ManyToOne(() => Showtime, { eager: true })
   @JoinColumn({ name: 'showtimeId' })
   showtime: Showtime;
+
+  // --- Other properties --------------------------------------
 
   @Column()
   userId: string;
