+++ src/movies/movies.service.ts
@@ -11,18 +11,6 @@ import { DTO_movie_update } from './DTO/update-movie.DTO';
 // requirement 2.1 done
 // requirement 3.1 done (ID validations)
 
-function validate_ID(id: number): void {
-    if (!id || id <= 0) {
-      throw new BadRequestException('Invalid movie ID provided.');
-    }
-  }
-  
-  function validate_DTO(dto: any, errorMessage: string): void {
-    if (!dto || Object.keys(dto).length === 0) {
-      throw new BadRequestException(errorMessage);
-    }
-  }
-
 @Injectable()
 export class Service_Movies 
 {
@@ -38,8 +26,6 @@ export class Service_Movies
 
     async create(create_DTO: DTO_movie_create): Promise<Movie> 
     {
-        validate_DTO(create_DTO, 'Movie data must not be empty.');
-
         return this.movie_repo.save(create_DTO);
     }
     
@@ -56,7 +42,6 @@ export class Service_Movies
 
     async update_by_title(title: string, update_DTO: DTO_movie_update): Promise<Movie> {
 
-        validate_DTO(update_DTO, "data must not be empty")
         const movie = await this.find_by_title(title);
         const updated = { ...movie, ...update_DTO };
 
@@ -73,13 +58,10 @@ export class Service_Movies
 
     async find_by_id(ID: number): Promise<Movie> 
     {
-        validate_ID(ID);
         const movie = await this.movie_repo.findOne({ where: { id: ID } });
 
         if (!movie) 
-        {
             throw new NotFoundException(`movie (ID ${ID}) not found`);
-        }
 
         return movie;
     }
@@ -87,7 +69,9 @@ export class Service_Movies
     async find_by_title(title: string): Promise<Movie> {
 
         const movie = await this.movie_repo.findOne({ where: { title } });
-            if (!movie) throw new NotFoundException(`Movie "${title}" not found.`);
+        
+        if (!movie) 
+            throw new NotFoundException(`movie (TITLE ${title}) not found`);
 
         return movie;
       }
