+++ src/showtimes/showtime.entity.ts
@@ -2,23 +2,28 @@
 import { Entity, Column, PrimaryGeneratedColumn, ManyToOne, JoinColumn } from 'typeorm';
 import { Movie } from '../movies/movie.entity';
 
+
+// --------- 2.2.1 requirements 
+
 @Entity()
 export class Showtime 
 {    
 
-    // -------- ID ----------------------------------------------------------
+    // -------- Primary Key ----------------------------------------------------------
 
     @PrimaryGeneratedColumn()
     id: number;
 
-    // -------- 2.2.1 requirements ------------------------------------------
+    // -------- Foreign Key ------------------------------------------
+
+    @Column()
+    movieId: number;
 
     @ManyToOne(() => Movie, { eager: true })
     @JoinColumn({ name: 'movieId' })
     movie: Movie;
 
-    @Column()
-    movieId: number;
+    // -------- Other Columns  ------------------------------------------
 
     @Column()
     theater: string;
