+++ test/jest-e2e.json
@@ -1,9 +1,9 @@
 {
-  "globalSetup" : "./test-setup.ts",
+  "globalSetup" : "./test/test-setup.ts",
   "moduleFileExtensions": ["js", "json", "ts"],
-  "rootDir": ".",
+  "rootDir": "..",
   "testEnvironment": "node",
-  "testRegex": ".e2e-spec.ts$",
+  "testRegex": "\\.e2e-spec\\.ts$",
   "transform": {
     "^.+\\.(t|j)s$": "ts-jest"
   }
